Fix: 고객관리 게시글 본인만 수정 및 삭제 가능하도록 구현

+++ static/js/support.js
@@ -1,30 +1,29 @@
 history.scrollRestoration = "manual"
 
 $(document).ready(function () {
-    get_list();
+    getList();
 });
 
-function toggle_writing() {
+function toggleWriting() {
     $("#sumbit_toggle").toggle();
     $("#input_text").val('');
     $("#input_name").val('');
     $("#input_title").val('');
 }
 
-function toggle_myid(num) {
+function toggleMyId(num) {
     let buildID = '#contentsbox-' + num;
     $(buildID).toggle();
 }
 
-function submit_content() {
-    let newName = $("#input_name").val();
+function submitContent() {
     let newTitle = $("#input_title").val();
     let newText = $("#input_text").val();
-    let data = {"username": newName, "post_title": newTitle, "post_content": newText};
-    if (newName != null && newTitle != null && newText != null) {
+    let data = {"post_title": newTitle, "post_content": newText};
+    if (newTitle != null && newText != null) {
         $.ajax({
             type: "POST",
-            url: `/api/supports/`,
+            url: `/api/user/supports/`,
             contentType: "application/json",
             data: JSON.stringify(data),
             success: function (response) {
@@ -34,45 +33,62 @@ function submit_content() {
     }
 }
 
-function edit_contents(num) {
-    let setContentsID = '#contents-' + num;
+function editContent(contentIdx) {
+    let getToken = localStorage.getItem('IllllIlIII_hid');
+    let setContentsID = '#contents-' + contentIdx;
+    let setUsername = '#username-' + contentIdx;
+    let getUsername = $(setUsername).text();
 
-    if ($('.edit_btn').html() == "수정") { //수정 시작
-        $(setContentsID).attr("readonly", false);
-        $('.edit_btn').html('저장');
-    } else { //저장버튼 클릭 후
-        $('.edit_btn').html('수정'); //저장 버튼 클릭시 수정으로 글씨 변경 및 서버 전송
-        $(setContentsID).attr("readonly", true);
+    if(getToken==getUsername){
+        if ($('.edit_btn').html() == "수정") { //수정 시작
+            alert('수정을 시작합니다');
+            $(setContentsID).attr("readonly", false);
+            $('.edit_btn').html('저장');
+        } else { //저장버튼 클릭 후
+            $('.edit_btn').html('수정'); //저장 버튼 클릭 => 수정으로 글씨 변경 및 서버 전송
+            $(setContentsID).attr("readonly", true);
+            let contents = $(setContentsID).val();
+            let data = {"post_content": contents};
+            console.log(data)
+            $.ajax({
+                type: "PUT",
+                url: `/api/user/supports/${contentIdx}`,
+                contentType: "application/json",
+                data: JSON.stringify(data),
+                success: function (response) {
+                    alert('수정 완료');
+                }
+            });
+        }
+    }else{
+        alert("권한이 없습니다.")
+    }
 
-        let contents = $(setContentsID).val();
-        let data = {"post_content": contents};
-        console.log(data)
+
+}
+
+function deleteContent(contentIdx){
+    let getToken = localStorage.getItem('IllllIlIII_hid');
+    let setUsername = '#username-' + contentIdx;
+    let getUsername = $(setUsername).text();
+
+    if(getToken==getUsername){
+
         $.ajax({
-            type: "PUT",
-            url: `/api/supports/${num}`,
+            type: "DELETE",
+            url: `/api/user/supports/${contentIdx}`,
             contentType: "application/json",
-            data: JSON.stringify(data),
             success: function (response) {
-                alert('수정 완료');
+                window.location.reload()
+                alert("삭제완료")
             }
         });
+    }else{
+        alert("권한이 없습니다.")
     }
 }
 
-function delete_content(contentIdx){
-    console.log();
-    $.ajax({
-        type: "DELETE",
-        url: `/api/supports/${contentIdx}`,
-        contentType: "application/json",
-        success: function (response) {
-            window.location.reload()
-            alert("삭제완료")
-        }
-    });
-}
-
-function get_list() {
+function getList() {
     $.ajax({
         url: '/api/supports',
         type: 'GET',
@@ -92,9 +108,9 @@ function get_list() {
                             <tr class="tr_title" id="open_it">
                                 <th scope="row">${idx}</th>
                                 <td class="name">
-                                    <a onclick="toggle_myid(${idx})"> ${getTitle} </a>
+                                    <a onclick="toggleMyId(${idx})"> ${getTitle} </a>
                                 </td>
-                                <td>${getUsername}</td>
+                                <td id="username-${idx}">${getUsername}</td>
                                 <td>${createdYYMMDD}</td>
                             </tr>
                             <tr id="contentsbox-${idx}" class="content_box" style="display: none;">
@@ -102,7 +118,7 @@ function get_list() {
                                     <div >
                                         <div class="content_btn_div" style="display: flex; justify-items: right;">
                                             <div class="content_btn_sub_div" style="margin-left: auto; padding-bottom: 10px;">
-                                                <button onclick="delete_content(${idx})" type="button" class="submit_btn btn btn-light">게시글 삭제하기</button>
+                                                <button onclick="deleteContent(${idx})" type="button" class="submit_btn btn btn-light">게시글 삭제하기</button>
                                             </div>
                                         </div>
                                         <div class="input-group mb-3" >
@@ -118,8 +134,8 @@ function get_list() {
 
                                         <div class="content_btn_div" style="display: flex; justify-items: right;">
                                             <div class="content_btn_sub_div" style="margin-left: auto;">
-                                                <button onclick="edit_contents(${idx})" type="button" class="edit_btn btn btn-light">수정</button>
-                                                <button onclick="toggle_myid(${idx})" type="button" class="submit_btn btn btn-light">닫기</button>
+                                                <button onclick="editContent(${idx})" type="button" class="edit_btn btn btn-light">수정</button>
+                                                <button onclick="toggleMyId(${idx})" type="button" class="submit_btn btn btn-light">닫기</button>
                                             </div>
                                         </div>
